requirements.txt
colorama==0.4.6
pyotp==2.9.0
python-dotenv==1.0.1
